requirements.txt
black==21.11b1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.2
pyTelegramBotAPI==4.2.1
python-dotenv==0.19.2
regex==2021.11.10
requests==2.26.0
tomli==1.2.2
typing-extensions==4.0.1
urllib3==1.26.7
